AI PATCH 1.1
Win scores are now scaled by the depth at which they are found. E.g. if the bot can win in 1 move the score is 10000 * (depth + 1). If the bot can also win in 6 moves, the score is 10000 * ((depth - 5) + 1) [assuming depth > 6]. This prevents the algorithm from "teasing" wins when it is able to win on its current move.

This is due to the way my algorithm functions. E.g. with a depth of 4, the minmax will begin searching a tree, starting with the 0th column (yes they are zero indexed). If it finds a win within these 4 moves, it returns a very large score in attempt to bias the algorithm towards this choice. This issue in this is that for any of the other columns to be chosen, they must have a score > the current best score (column 0), regardless of how many moves are expelled to achieve a win. Therefore scaling the values by depth (number of moves effectively) corrects this bias.